move pubsub into server
